Some refactoring

+++ src/base/tl/allocator.h
@@ -10,7 +10,7 @@ public:
 	static T *alloc() { return new T; }
 	static void free(T *p) { delete p; }
 
-	static T *alloc_array(int size) { return new T [size]; }
+	static T *alloc_array(size_t size) { return new T [size]; }
 	static void free_array(T *p) { delete [] p; }
 };
